refactor(workflow): extract WorkflowApp component and improve resource cleanup
- Extract WorkflowApp component from index.tsx for better code organization
- Add dispose handler in workflow.ts for proper resource cleanup
- Wrap workflow app with AppWrapper component
- Update script extension from .tsx to .js in workflow.html

This change improves component modularity and memory management while maintaining
existing functionality.
