+++ vscode/webviews/workflow.html
@@ -11,6 +11,6 @@
     </head>
     <body>
         <div id="root"></div>
-        <script type="module" src="workflow/index.tsx"></script>
+        <script type="module" src="workflow/index.js"></script>
     </body>
 </html>
