+++ vscode/webviews/workflow/index.tsx
@@ -1,96 +1,14 @@
-import {
-    Background,
-    Controls,
-    type EdgeChange,
-    type NodeChange,
-    ReactFlow,
-    addEdge,
-    applyEdgeChanges,
-    applyNodeChanges,
-} from '@xyflow/react'
-import React, { useState, useCallback } from 'react'
+import React from 'react'
 import ReactDOM from 'react-dom/client'
-import '@xyflow/react/dist/style.css'
 
-const initialNodes = [
-    {
-        id: '1',
-        data: { label: 'Git Diff' },
-        position: { x: 0, y: 0 },
-        type: 'input',
-    },
-    {
-        id: '2',
-        data: { label: 'Cody Generate Commit Message' },
-        position: { x: 100, y: 100 },
-    },
-    {
-        id: '3',
-        data: { label: 'Git Commit' },
-        position: { x: 100, y: 100 },
-    },
-]
-
-const initialEdges = [
-    { id: '1-2', source: '1', target: '2', type: 'step' },
-    { id: '2-3', source: '2', target: '3', type: 'step' },
-]
-interface WorkflowEditorProps {
-    count?: number
-}
-
-const WorkflowEditor: React.FC<WorkflowEditorProps> = () => {
-    const [nodes, setNodes] = useState(initialNodes)
-    const [edges, setEdges] = useState(initialEdges)
-
-    const onNodesChange = useCallback(
-        (
-            changes: NodeChange<
-                | {
-                      id: string
-                      data: { label: string }
-                      position: { x: number; y: number }
-                      type: string
-                  }
-                | {
-                      id: string
-                      data: { label: string }
-                      position: { x: number; y: number }
-                      type?: undefined
-                  }
-            >[]
-        ) => setNodes(nds => applyNodeChanges(changes, nds) as typeof initialNodes),
-        []
-    )
-
-    const onEdgesChange = useCallback(
-        (changes: EdgeChange<{ id: string; source: string; target: string; type: string }>[]) =>
-            setEdges(eds => applyEdgeChanges(changes, eds)),
-        []
-    )
-
-    const onConnect = useCallback((params: any) => setEdges(eds => addEdge(params, eds)), [])
-
-    return (
-        <div style={{ width: '100%', height: '100vh' }}>
-            <ReactFlow
-                colorMode="light"
-                nodes={nodes}
-                edges={edges}
-                onNodesChange={onNodesChange}
-                onEdgesChange={onEdgesChange}
-                onConnect={onConnect}
-                fitView
-            >
-                <Background />
-                <Controls />
-            </ReactFlow>
-        </div>
-    )
-}
+import '../index.css'
+import { AppWrapper } from '../AppWrapper'
+import { WorkflowApp } from './WorkflowApp'
 
 ReactDOM.createRoot(document.querySelector('#root') as HTMLElement).render(
     <React.StrictMode>
-        <WorkflowEditor />
+        <AppWrapper>
+            <WorkflowApp />
+        </AppWrapper>
     </React.StrictMode>
 )
